+++ lib/phoenix_kit_crm/web/companies_live.ex
@@ -8,7 +8,7 @@ defmodule PhoenixKitCRM.Web.CompaniesLive do
   alias PhoenixKitCRM.{Activity, Companies, PartyRoles, Paths}
   alias PhoenixKitCRM.Schemas.Company
 
-  @role_filters ~w(supplier customer)
+  @role_filters ~w(supplier customer manufacturer partner)
   @page_size 25
 
   @impl true
@@ -133,6 +133,20 @@ defmodule PhoenixKitCRM.Web.CompaniesLive do
             class={["tab", @filter == "customer" && "tab-active"]}
           >
             {gettext("Customers")}
+          </.link>
+          <.link
+            patch={companies_path(assigns, filter: "manufacturer", page: 1)}
+            role="tab"
+            class={["tab", @filter == "manufacturer" && "tab-active"]}
+          >
+            {gettext("Manufacturers")}
+          </.link>
+          <.link
+            patch={companies_path(assigns, filter: "partner", page: 1)}
+            role="tab"
+            class={["tab", @filter == "partner" && "tab-active"]}
+          >
+            {gettext("Partners")}
           </.link>
           <.link
             :if={@trashed_count > 0 or @filter == "trashed"}